added CXX_FLAG "-Wuseless-cast" & adjusted warns

+++ src/cc/include/swcdb/core/Time.h
@@ -29,14 +29,14 @@ void checkings();
 SWC_CAN_INLINE
 extern
 int64_t now_ms() {
-  return (int64_t)std::chrono::duration_cast<std::chrono::milliseconds>(
+  return std::chrono::duration_cast<std::chrono::milliseconds>(
     std::chrono::system_clock::now().time_since_epoch()).count();
 }
 
 SWC_CAN_INLINE
 extern
 int64_t now_ns() {
-  return (int64_t)std::chrono::duration_cast<std::chrono::nanoseconds>(
+  return std::chrono::duration_cast<std::chrono::nanoseconds>(
     std::chrono::system_clock::now().time_since_epoch()).count();
 }
